Badluck Initial Patch
This patch contains a broad suite of changes targeted at removing RNG from the game in such a way as to megativley affect the player in basically every cirumstance. The content of the game is mostly unchanged (except for Steven). , as the design goals was for everything that happens to be technically possible within a base emerald.